+++ changelog.md
@@ -5,3 +5,4 @@ Initial release.
 
 
 patch: fixes issue where in Dwcom would not recognize that speechInterrupt was set to false.
+patch: Makes speach output say left/joined channel root rather than left/joined channel the root channel.